TEST-52 review 삭제 방식 POST

+++ foreign/views.py
@@ -66,7 +66,6 @@ def univ_search(request):
                     last_alpha = this_alpha
                 else:
                     univ_dict[this_alpha].append(univ)
-            print(len(univ_dict['A']))
         if len(univ_dict['A']) == 0:  # A인 대학이 없을 때 A출력 제거
             del(univ_dict['A'])
 
@@ -337,8 +336,9 @@ def review_create(request, foreign_id, post=None):
 
 
 def review_delete(request, pk, foreign_id):
-    review = get_object_or_404(Post, pk=pk)
-    review.delete()
+    if request.method == 'POST':
+        review = get_object_or_404(Post, pk=pk)
+        review.delete()
     return redirect('foreign:review_list', foreign_id)
